Lots of changes.  Principally, support for library specs files is now in place.  Eclipse, though, does not like more than one java-like file in the same classpath, so for now, at least, spec files have to be in a separate project and at the moment you have to set that up manually with the EditSpecs dialog under the JML menu.  But there is an API for getting at the specs for given types now that I hope is reasonably stable.  Added some of the specs from ESCTools as a start.
The handling of types and the whole parsing and typechecking need
an overhaul.

There is also a Show Specs dialog that is helpful in debugging and
in looking at the specs for various Java elements.

Lots of other little additions, cleanups, refactorings, total redesigns
etc.  The question of whether spec files should be java-like or not is
still open.  Eclipse has nuisances either way.

Also - still hoping for help/guidance from Patrice and Barry 
on the Eclipse parser and perhaps other stuff.

Finally, for prospective developers out there, note the new
setup file - the Import mechanism is a much easier way to
setup to use the SVN distribution of JML3 than was my old
instructions.