A point a which tests pass, but not a lot of actual default behavior change. However, an ANTLR parser for Java and JML is now present, along with bug fixes, a lot more tests, and some reorganization of code and tests.
At this point I expect to do only essential corrections and embellishments in the parsers
(the custom and the antlr) parser, while I wait to see what Chalin&James are doing in JML4,
as I would really rather make more use of the internals of Eclipse for parsing and ASTs.

Also, I'd like to make use of Boysen's work on using Java annotations, at least for modifiers.

In the meantime, I'll proceed with implementing generic checks on the JML AST that any parser
will need, put in the beginning of runtime checking, since that will enable type and name resolution
without having to reimplement all that complicated logic.

Note that because of the antlr grammar, you will need to add the ANTLR (2.7.6) plugin from
http://antlreclipse.sourceforge.net/updates