[esc4] initial support for numeric quantifiers        fixes for conditional operator, &&
CfgSubstitutionVisitor:
new class.  replaces CfgVariables with value given in map passed to constructor.

PassifyVisitor:
added initial (experimental) support for inlining a method call's formal parameters with the bound value, controlled with a boolean.
wrapped stmtsFromExprs as a separate static class and added a toString() to help with debuging.
fixed bug with handling of conditional expressions containing method calls with preconditions.
getNewAssignments renamed to handleNewAssignmentsAndAsserts

TargetGatheringVisitor:
added a toString() to help with debuging.

VarToOldVisitor:
fixed bug with handling of conditionalexpressions that caused body to be a copy of range

ProverVisitor (and subclasses):
added node type to support an "n-ary and"
added initial support for numeric quantifiers (product & sum)

Cvc3Visitor:
if no input received from process's input stream, read from it's error stream.

IsabelleAdaptor:
added UBP, first needed to support numeric quantifiers

VcConditionalExpression:
fixed translation of "a && b" to "a & (a -> b)" when splitting VC into sub-VCs.

VcProgram:
getAsImplications now returns an n-ary conjunction implying the last one
(instead of a -> b -> c it now returns (/\ a b) -> c)

tests:
added unit tests for numeric quantifiers and updated those that changed because of the change in handling of &&