[jml4] JML4 Quantifiers Bugfix Patch submitted by Dan Zimmerman
This patch fixes two issues with quantifier parsing and resolving: 1) the
"degenerate" form (\forall int a; ; a < 0); - that is, the quantifier with
an empty range with semicolon (as opposed to an omitted range and
correspondingly omitted semicolon) - did not parse properly, resulting in
odd errors; and 2) a null pointer exception was generated under certain
circumstances when resolution of variables outside the quantifier's own
scope failed.

In addition to very minor grammar changes and changes to the
JmlQuantifiedExpression class, two unit tests were added to test for these
issues.

-Dan Zimmerman, UW Tacoma, 2008-07-08