[jml4] Assignable-Clause Patch Part 2.  Submitted by Jooyong Lee.
<Overview>

This patch principally deals with error messages regarding assignable
clauses. More specifically, the following were added or fixed:

1. Position values kept in JmlName objects were adjusted. The original
values, in general, pointed at the next position of the expected one.

2. Added code for error reporting.

====================================
<Changes>


[Parser.java]

1. The original scanner has a problem to set as the values of
sourceStart and sourceEnd at method consumeStoreRefSuffixWildcardOrAll
one larger values than actual ones. This problem occurs due to two
reasons. The first reason is that the production rule corresponding to
the consumeStoreRefSuffixWildcardOrAll method is consumed only when
the token next to * is scanned. This position next to * is saved in
field startPosition. The second reason is that field currentPosition
used in the original consumeStoreRefSuffixWildcardOrAll method is one
larger than the value of startPosition. In other words, the value of
currentPosition points at the next next token of *. This problem is
fixed in the submitted patch.

2. Similarly to the above, the
consumeStoreRefNameOrSuffixAsSuperOrThis method was changed to set
variable sourceEnd with a right value.

[JmlName.java]

1. Corrected the position field value.

[ProblemReporter.java]

1. Added the illegalUsageOfThis method that can be used to report an
illegal use of 'this', e.g., this.this.

2. Added the invalidSuperAsStoreRef and invalidThisAsStoreRef methods
that can be used to report an error when 'super' or 'this'
respectively is erroneously declared as assignable.

[IProblem.java, messages.properties]

1. Added fields and message entries according to the changes described
above.

[JmlStoreRefExpression.java]

1. Replaced RuntimeException-throw statements with appropriate
error-reporting methods if they are reachable. We used four types of
reporting methods: the three methods described above and the
illegalUsageOfWildcard method to report an error when a wildcard is
followed by more elements, e.g., a.*.b. (Though this use of wildcards
in the middle of an expression could be considered legal, the
resolution of it seems expensive whereas it appears to be useful only
in a few restrictive cases.)

2. Some of the RuntimeException-throw statements are unreachable under
the current grammar rules. In this case, we left the statement as an
alarm trigger for possible future grammar changes along with a comment
explaining this intention. (prj: these throws were replaced with asserts).

3. Refined method resolve to report an error when this or super by
themselves are used as a store-ref expression.

4. Simplified the first switch statement of method resolveNameRef
since the sort of singleStoreRefName is always JmlName.SORT_IDENT. The
added JML specification explains this situation.

[JmlStoreRefKeyword.java]

1. Changed the visibility of field code into public so that it can be
known which type of keyword is used.

[java.g]

1. Corrected the redableName of storeRefName for more precise error
reporting.

====================================
<Test>

[AssignableTest.java]

1. Added test cases for error reporting.

2. Corrected error-position information of
test_Neg_UnresolvableClauseTest.


August 23, 2008
-Jooyong Lee