JML4 Commit Notes - 2008-12-10 WHAT is in this commit?
    * Essentially added ability to programmatically control the compiler phases applied to JML AST nodes. The phases are defined in JmlConstants.java. Eventually, this might be turned into an Eclipse preference.
    * Added (or enhanced) ability to determine JML modifiers for fields, constructors and methods from Java annotations.

DESIGN

There is not much to say about the control of the compiler phases; look at the code diffs for this commit, starting with JmlConstants.

JML Modifiers. JML modifiers, once parsed, are stored internally in an AST as Java 5 annotations, even when parsing Java 1.4 code! As is typical in compilers, ordinary Java modifiers are represented as bits in a field of integral type. Similarly, to support JML modifiers we have the following extra fields and methods:

    private long jmlModifiers = 0;
    private boolean jmlModifiersHaveBeenInit = false;

    public void initJmlModifiersFromAnnotations() {
      jmlModifiers |= JmlModifier.getFromAnnotations(this.annotations);
      this.jmlModifiersHaveBeenInit = true;
    }

    public boolean isPure() {
    	if (!this.jmlModifiersHaveBeenInit)
    		initJmlModifiersFromAnnotations();
    	return JmlModifier.isPure(this.jmlModifiers);
    }
    ...

This is only an intermediate commit. It will make sense collect all JML modifier responsibilities into JmlModifier.

DETAILS

    * JmlConstants.java: added constants to control how many JML4 compiler phases are applied to JML constructs.
    * Various AST methods updated to help control compilation phases for JML constructs and/or proper support for JML modifiers.