[http://jmlspecs.svn.sourceforge.net/viewvc/jmlspecs/jml4/trunk/org.eclipse.jdt.core/notes/commit/2008-12-16.html?revision=1186]
Commit 1186 - start of the experimental merge of JML5 into JML4

As of this commit, JML4 should be run with a Java 6 or higher JVM (previously it was Java 5), and JML projects will need to be set to a Java 6 compliance level or higher. JML projects must also have org.jmlspecs.annotation in their build path.

Prelude

This is checkpoint commit, the first of a series of commits, adding experimental support for JML5 (i.e. Java 5 annotations for JML) to JML4. JML5 was developed by Kristina Taylor in her master's thesis under the supervision of Gary Leavens:

    Taylor, Kristina B. (2008) A Specification Language Design for the Java Modeling Language (JML) Using Java 5 Annotations . Masters thesis, Computer Science, Iowa State University.

Another closely related article is

    Taylor, Kristina B., Rieken, Johannes and Leavens, Gary T. (2008) Adapting the Java Modeling Language for Java 5 Annotations. Technical Report 08-06, Computer Science, Iowa State University.

which also describes Modern Jass and how it supports a subset of the Java 5 annotations for JML.

WHAT is in this commit?

This commit has JML4 switch to using JML5 annotations to represent JML modifiers (e.g. @Nullable, @SpecPublic) rather than our previous homegrown org.jmlspecs.jml4.runtime.annotation. This switch caused some of the code resolution for JML set statements to no longer work, so that has been adjusted as well.

DETAILS

Selected details:

    * org.jmlspecs.annotation:
          o Added Java 5 annotations for JML modifiers that were missing; i.e. the math modes and instance modifiers.
          o @Ghost and @Model have been given a default value so that they can be used as simple modifiers (rather than "declarator" annotations).
    * org.eclipse.jdt.core:
          o The main changes in this package are to JmlModifier and TypeConstants to support constants (mainly char arrays) to hold the simple and fully qualified names of the JML5 annotations.
          o Aligned processing of the ghost modifier in JmlSetStatement as well as VariableBinding, FieldBinding and LocalVariableBinding to make use of the @Ghost annotation rather than a JML specific field in the binding classes named isGhost. (VariableBinding and FieldBinding now match their Eclipse JDT counterparts; i.e. all JML4 specific changes have been eliminated.)
    * org.jmlspecs.eclipse.jdt.core.tests:
          o Test files updated to make use of org.jmlspecs.annotation.
          o New Jml5BasicTests.java.