I add two axioms that specify the hexademical floating point literals that MAX_VALUE and MIN_VALUE are equal to.
Then I change the isFinite method so that it is true if and only if
the input is <= Max_VALUE and >= -MAX_VALUE.

Note that Java 8 doubles that satisfy the above two inequalities are
finite by definition.

This new version of isFinite helps OpenJML verify behavior of Math
methods on finite inputs by avoiding the isInfinite method, which
involves comparisons with POSITIVE_INFINITY and NEGATIVE_INFINITY
that OpenJML cannot verify properly due to outstanding issues with
infinity.